Add simulation interface and enable backend as a web-worker.
worker.js contains a launcher for the simulation backend meant
to be run inside a web-worker. It creates a core and registers
a controller to handle messages sent to the worker.

LogikSim.Backend.Simulation handles launching such a web-worker
running worker.js and communicating with it. A Simulation instance
acts as a proxy between the backend Controller in the web-worker
and the front-end user. It exposes all available functionality as
normal function calls and allows registration of handlers for
reacting to backend events.