Add Component and component library as well as edge events.
This commit adds the basic structures for a type 'Component'
and the component library. A (circuit) component is the basic
building block of the logic simulation. The component has a
type specific logic function it uses to calculate outputs for
corresponding inputs applied to it.

State changes on the component are realized via a clocked two
stage interface. Every cycle input changes for one or more input
changes for the component have to be processed the edge function
has to be called for each of them. For every clock cycle in which
edge calls were performed a corresponding clock call enables the
component to react to these changes in terms of follow-up events
which are scheduled in the core.

This edges then clock behavior is wrapped in the implementation
of so called Edge events which represent rising or falling edges
on a specific component input at a specific time. Edges use the
component id as group identifier which enables them to use the
last_in_group feature of the core to reliable trigger a clock
event only once per clock cycle in which the Component received
edge events.

For convinience components also use so called OutEdges which can
be seen as internal signals traveling through the logic gates.
OutEdges are generated for future output port changes on a component.
This reason for this indirection is the interactive nature of
this simulation. If the simulation structure was fixed and we were
only interested in some final output OutEdges would be redundant
as we could directly schedule Edges with the internal delay applied
on component inputs connected to our component's outputs.
However as we want the component to propagate state changes to
the front-end when its output pins change and we cannot guarantee
a component that was connected to a specific output when our input
changed is still connected there this two stage scheme is employed.
OutEdges on a component should never be schedule by anything except
the component itself.

Component behavior is derived from so-called templates which can
be expressed in terms of JSON serializable objects. The component
library stores such templates and allows Component instantiation
from a previously registered template. Templates are identified
by their type which ideally should be a globally unique identifier.

component_templates.json contains a first draft of how JSON
serialized descriptions that can be loaded by a component library
might look like.

This current commit lays most but not all of the groundwork for
the concepts described above. Later commits will expand on these
foundations so to get a full picture reviewing this commit on its
own might not be a good use of time.