Implement component state propagation functionality.
The basic structures for state propagation as well as the
interface components for state manipulation via the simulation
object are now implemented.

State propagation relies on Component property changes being
reliable propagated up from child components to their parents
until they reach a parent_root handled by the simulation
controller. There they are wrapped and transfered into the
front-end. The goal of state propagation is to ensure that
the front-end always (eventually, hopefully sooner than later)
has the complete set of component properties. As a components
full set of properties should be sufficient to re-create it
and derive its observable state this is a flexible and powerful
way to keep the UI informed of what is happening in the
simulation backend.

For proper testing we should first implement the synchronous
backend connectivity so we can actually debug the interaction.
Unfortunately JavaScript tooling is extremely weak when it
comes to debugging web-workers.