Current implementation can now do (very) basic simulation.
This commits fixes various remaining issues around the practical
implementation of the controller, simulation, web-worker etc.

Finding these issues bug was delayed by the insufficiency of JavaScript
debugging capabilities in web-workers. Especially in conjunction with
Karma. This was resolved by implementing the debugging feature
described in #3 . This enables backend simulation/controller interaction
without web-workers. This way normal JavaScript debugging methodology
can be used to debug the tests which greatly simplifies debugging.

With this commit the first very basic "simulation" could be performed
by the backend. test_simulation.js creates an AND gate connected to the
input of an OR gate. It then schedules rising edge events on the and
gate inputs and tests whether the resulting change in the AND output
cascades through the OR gate which should propagate it to the "frontend".

Fixes #3