Improved test coverage and documentation for existing interfaces.
The improved test coverage consist mainly solely of integration tests
again the Simulation interface. To make it easier to debug the simulation
interface tests they are now run twice in two distinct test suites. One
using the real-world web-worker backend and the other one not using
web-workers and configured to crash on errors. This should enable most
errors the testing can detect to provide a proper trace during test
execution instead of (in the best case) a message by our custom error
handling.

The documentation is a bit simplistic at this point but should be
sufficient to get a good idea of the functionality exposed on the
interfaces. It'll likely be best to expand it later on when we
actually programming against it than trying to nail it now. Otherwise
we might end up throwing away lots of work with later changes.