Add new integration test and fix issues found by it
This patch adds a new simulation test which builds a half-adder
and checks its behaviour against a corresponding truth table. The
adder consists of interconnects, an and and a xor.

Doing so revealed that the Interconnect logic function was faulty
due to unexpected behaviour of the array map function with present
but undefined elements.

Also the interconnect is much easier to use when it automatically
adds outputs as needed. For enable this components now will add
the needed number of output ports on-the-fly when asked to connect
to a yet non-existent port.

The simulation _gen_element_uid function incorrectly gave the first
element the id 0. As 0 cannot be properly negated it breaks assumptions
w.r.t to that in the out edge handling.

Made the delay parameter in the connect function on the simulation interface
optional with default 0 to simplify common use.

Changed interface and component templates to consistently use boolean
instead of integer values for input and output states.