Rewrote run_async_truth_table_simulation to be more robust
Using the new steady state signaling as well as the ability
to stop and restart simulations the run_async_truth_table_simulation
can be implemented in a much more robust and performant way.
The new approach simply applies to inputs from one row of the
truth table, starts the simulation and then waits until the
core reports a steady simulation state to check the results.
To switch to the next row of the truth table the simulation
is stopped and the process repeated. This way the simulation
can run as fast as possible and there is no danger of checking
a transient state like before.